Fix bug in CreateUser() call.
New accounts seemed not to have the correct password_hash_function set. Checking
against
https://developers.google.com/google-apps/provisioning/#creating_a_user_account,
it seems I have the wrong parameter order. I'm passing the
password_hash_function as the 5th param. According to that documentation,
the 5th param is a boolean to indicate whether or not the account is suspended.

Passing a boolean as the 5th parameter threw an exception.

So I'm just naming the params. It works.